// dashboard/src/utils/hasura-api/generated/schemas/logicalModelSelectPermissionItem.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { LogicalModelSelectPermission } from './logicalModelSelectPermission';

export interface LogicalModelSelectPermissionItem {
  role: string;
  permission: LogicalModelSelectPermission;
}